requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==4.2
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environ==1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.8
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kavenegar==1.1.2
kombu==5.4.0
model-bakery==1.20.0
openai==1.61.1
packaging==24.2
prompt_toolkit==3.0.47
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==4.2
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
environ==1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.8
inflection==0.5.1
jalali_core==1.0.0
jdatetime==5.0.0
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kavenegar==1.1.2
kombu==5.4.0
model-bakery==1.20.0
openai==1.61.1
packaging==24.2
phonenumbers==9.0.3
prompt_toolkit==3.0.47
psycopg==3.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
